requirements.txt
streamlit==1.32.0
pandas==2.1.4
google-generativeai==0.3.2
sentence-transformers
faiss-cpu
numpy==1.26.0
scikit-learn==1.3.2